% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspiration-package.R
\docType{package}
\name{inspiration-package}
\alias{inspiration-package}
\title{inspiration: This package helps to normalize data eaisly using log transformation.}
\description{
This package provides a quick log transformation function to normalize the data
}
\author{
\strong{Maintainer}: Esin Gurcan \email{esiingurcaan@gmail.com}

}
\keyword{internal}
