% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspiration-package.R
\name{inspiration}
\alias{inspiration}
\title{inspiration: a package created for the week 7's assignment}
\description{
This oackage is an exercise for me to learn how to work with the data files
in package using R Studio. Therefore this package includes one of my data
fies from pervious publication
}
\author{
Esin Gürcan, McMaster PNB
}
