% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{A function for normalizing the data using log transformation}
\usage{
normalize(x, base = 10)
}
\arguments{
\item{a}{A is a data column that you want to do log transformation}
}
\value{
it returns to the log transformed values
}
\description{
It is a simple log transformation function to normalize the data
}
\examples{
 data <- sample(1:100, 5)
 normalize(data)
}
